Decoded list fix
Fixed the decoded list from growing exponentially, by moving the processSerial function to the decode function. This significantly reduced the audio delay. There still is a slight audio delay that gets output about a second after rekeying. For this fix, the processSerial function in the read thread was commented out and this casued a delay in receiving some of the control signals from the dongle.